Add CtrvStateCovariance and VehicleObject
The CtrvStateCovariance is a type alias for an Eigen matrix. It would be nice
to have a units-based type, but this is currently a hard limitation for all
open source linear algebra libraries. The VehicleObject is a type alias for
a DetectedObject with a CTRV motion model.

dependencies.cmake
* add Eigen3

include/cooperative_perception/ctrv_model.hpp
* add CtrvStateCovariance type alias

include/cooperative_perception/detected_object.hpp
* add VehicleObject type alias

src/CMakeLists.txt
* link to Eigen3