Update VS Code workspace file associations
The workspace settings allowed file associations to be automatically
added to the workspace settings, but this meant any standard header file
included in the project would be added to this list (since header files
have no extension). This is unnecessary, so automatic file associations
have been disabled. Existing associations have also been removed because
they kind of snuck in.

.vscode/settings.json
* disable automatic file associations
* remove existing file associations