Add std::hash specialization for CtrvState
A hash mechanism and equality operator must be provided to use
user-defined types in unordered containers. Previously, the boost::hash
specialization for CtrvState was used, but this required passing it as a
template argument whenever CtrvState was used in an unordered container.
This was a bit annoying, so it was replaced with an alternative where
the std::hash specialization for CtrvState is injected into the std
namespace.

include/cooperative_perception/ctrv_model.hpp
* remove hash_value function because it is no longer needed
* add std::hash specialization for CtrvState

include/cooperative_perception/unscented_transform.hpp
* remove boost::hash template argument

src/ctrv_model.cpp
* remove hash_value function implementation

src/unscented_transform.cpp
* remove boost::hash template arguments

test/test_unscented_transform.cpp
* remove boost::hash template arguments