11 angle type (#17)
Description
This PR contains the implementation of a Angle type that will be used within the model types, e.g. CtrvState, CtraState, to better manage sampling algorithms. This type assumes that angles take values from [0, 2pi). The type has several arithmetic operators that will support downstream operations, such as weighted means.

Related Issue
New angle class: #11

Motivation and Context
This problem solves the issue of sampling angles when near the boundary of 0 and 2pi.

How Has This Been Tested?
The implementation was validated via pen and paper and Python scripts. Formal unit tests for the operators were created.

* Created Angle class and test file stubs

The basic imports and boilerplate for defining and testing angles were added to the repo.

* Angle constructor and accessors implemented

Created the Angle constructor and its getter and setter pair.

* Angle class arithmetic fixes

Detected errors with the arithmetic on paper based on current test harness values.
Fixed the +=, -=, and *= operations.

* Revisions and initial test file

I updated the Angle.hpp to apply the recommendations in the review.
Also, I implemented the AngleArithmeticOperations test function.

* Added boolean equality operator

Created a boolean == operator to allow the direct comparison of two Angle objects.
Created a test for this new operator.

* Updates to Angle class

Revised some of the Angle class functionality:
+ Created exact == semantics using the underlying complex class
+ Created almostEqual function
+ Added EXPECT_NEAR macro to Angle test file.
+ Changed to using Boost.Math constants for test values in Angle's test file.

* Swapped in Boost Math constant

Swapped out the M_PI constant for Boost's math constant representation of PI.

* Comments updated

Added explanatory comments to the Angle class as well as its almostEqual implementation.