#28 Add track to object scoring
The detected object and track classes now have UUID fields because it was
needed for the map used in the scoring function. It would have been nice to
use references to the actual objects, but it was looking pretty ugly.

The scoring function uses the Euclidean distance metric. It would be useful to
modify the implementation to allow different metrics. Unit tests for the
scoring function have also been added.