Add support for object-to-track scoring (#31)
* #28 Add Euclidean and Mahalanobis distances

We can now calculate the Euclidean distance between two points and the
Mahalanobis distance between a point and a distribution. Each motion model
provides functions for calculating the distances, and the scoring.hpp file
provides object-to-track scoring versions.

* #28 Update Euclidean distance function signatures

They were creating copies of the CTRA and CTRV objects when they didn't need
to. Now they accept arguments as const refs.

* #28 Add track to object scoring

The detected object and track classes now have UUID fields because it was
needed for the map used in the scoring function. It would have been nice to
use references to the actual objects, but it was looking pretty ugly.

The scoring function uses the Euclidean distance metric. It would be useful to
modify the implementation to allow different metrics. Unit tests for the
scoring function have also been added.

* #28 Update spell check list with new words

* #28 Update distances to use only pose

The Euclidean and Mahalanobis distances used the full state vector, but this
makes it impossible to compare distances between entities with different
motion models. All motion models (so far) have elements for position (x, y)
and orientation (yaw). The distances now use only these three elements when
calculating the distances. Unit tests have been updated to reflect these
changes.

Co-authored-by: Patrick Martin <patrickmartin@fastmail.com>

* #28 Add Mahalanobis distance as scoring option

The score_tracks_and_objects function now accepts a distance metric argument,
which as of right now can be Euclidean distance of Mahalanobis distance.

---------

Co-authored-by: Patrick Martin <patrickmartin@fastmail.com>