Migrate dependencies to CPM
CPM is a package manager for CMake that allows us to bring in
project dependencies from source and build them alongside the
project. This allows us to better couple the project with its
dependencies. The project currently relies on an external package
manager (apt) to bring in Boost because building it from source
will take a long time. There is an enhancement in CPM open to
make things easier.