Update motion model type aliases
This commit fixes #61 by aliasing the motion model covariance types with
fixed-size Eigen matrices. While this is not the strongest restriction,
it should help with most issues. For future changes, a semantic tag
could be added to avoid clashes with motion models that have identically
sized state vectors but with different variables.