#72 Replace std::string Uuid with a class
The UUID represented by a string has been replaced by a proper Uuid
class to promote type safety and self-documentation. The unit tests
have been updated to accommodate the new interface. Other areas of the
library will need to be updated to more seamlessly work with the Uuid
class. For now, they use the underlying std::string type contained in
the Uuid class.