Add support for more detections than tracks (#94)
* #93 Add support for more detections than tracks

The gnn associator now pads the cost matrix with phantom tracks if there
are more detections than tracks. After the association step, any
associations with phantom tracks are discarded.

* #93 Add unit test

* #93 Fix bug in resizing the cost matrix

When adding the phantom tracks, the original cost matrix gets coppied.
However, it seems that dlib does not do a deep copy in the copy ctor.
Replaced that with the dlib::tmp function, which according to the
documentation, will do what we need. The unit tests always pass now.

* #93 Fix minor typo