#106 Fix making matrix square
There was some issue with the implementation for making cost matrices
squared. Instead of using dlib::tmp, we are now explicitly creating
a new empty matrix then populating the data manually. This operation
was also moved into its own function. We can refine the implementation
later if needed, but the separate function abstracts the current
implementation.