Remove `FixedThresholdTrackManager`'s `update_track` function type specifiers (#116)
* #115 Remove track manager update type specifier

The track manager's update_track function used the state_type and
state_covariance_type member aliases for the function parameter types,
but std::variant track types don't work with this. It's easier to
remove those type requirements and refactor later.

* #115 Update DynamicObject setter types

The set_* functions for the DynamicObject class had issues where the
DetectedObject's state_type got deduced to a different type than the
state parameter's type. This is a similar issue for all set_* functions
with DynamicObject.

* #115 Add getters for DynamicObject

* #115 Add std::variant setters for DynamicObject

This is for setting uuid and timestamp.

* #115 Fix std::variant setter implementations

The function signatures were all wrong, so this fixes that. The
to-be-set objects were passed by const ref, and the std::variant
wrappers were missing arguments.

* #115 More setter updates

* #115 Update variant setters

The std::visit call was failing to compile because not all elements of
the cartesian product of types were handled properly.

* #115 Mote setter updates

Previous attempts were failing to compile on downstream projects.

* #115 Add state and covariance copy functions

The get_state and get_state_covariance functions currently have
limitations when used with std::variant because their underlying calls
return different types. This means the getters will not work with
std::variant right now. As an alternative, we can make copy_* functions.

* #115 Fix copy_state and copy_state_covariance

The previous implementations caused deduction errors for std::variant.

* #115 Add update_track overload

* #115 Fix copy_state bug

The function was setting the destination state to itself instead of
the source's state.