Fix track manager's handling of confirmed tracks (#146)
* Add occurrence count getter

* Change how occurrence counts increase

* Fix missing include

* Fix typo

* Fix min/max function

Was using max when it should have been min.

* Remove demotion from confirmed

We're having issues with tracks being on the tentative/confirmed
boarder, which causes them to not always be published. Maybe we should
not demote confirmed tracks. Instead, we can remove them when they
fall below the removal threshold. This will still keep the promotion
threshold though.

* Remove commented code

It is no longer needed